requirements.txt
psycopg2==2.9.2
SQLAlchemy==1.4.28
pandas==1.3.4
fastapi==0.70.0
uvicorn==0.15.0
PyJWT==2.3.0
firebase_admin==5.2.0 
python-dotenv==0.19.2